requirements.txt
ultralytics==8.3.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.66.0
lxml>=5.2.0
PyYAML>=6.0
pillow>=10.0.0

# Gemini
google-generativeai>=0.8.0
python-dotenv>=1.0.1
